Reorganized housing models. Moved Service models from Context to Library. [#158039176]
