+++ src/Housing.Selection.Library/HousingModels/Batch.cs
