+++ src/Housing.Selection.Library/HousingModels/Room.cs
