+++ src/Housing.Selection.Library/HousingModels/User.cs
