+++ src/Housing.Selection.Library/ServiceHubModels/ApiUser.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 
-namespace Housing.Selection.Context.HttpRequests.ApiModels
+namespace Housing.Selection.Library.ServiceHubModels
+
 {
     public class ApiUser
     {
